Projects/E-Commerce/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
mysqlclient==2.2.4
stripe==7.11.0
python-dotenv==1.0.1
Pillow==10.1.0
django-cors-headers==4.3.1
django-filter==23.5
django-storages==1.14.2
boto3==1.34.34
gunicorn==21.2.0
whitenoise==6.6.0
dj-database-url==2.3.0
requests==2.31.0
python-slugify==8.0.1
django-crispy-forms==2.0
crispy-bootstrap5==0.7 